Title: 두 수의 거듭제곱, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 231231/두 수의 거듭제곱/two-numbers-of-squares.cpp
@@ -5,6 +5,6 @@ int main() {
     // 여기에 코드를 작성해주세요.
     int a,b;
     cin >> a >> b;
-    cout << pow(a,b);
+    cout << (int)pow(a,b);
     return 0;
 }
